moved parse examples back to parse_test.go
