+++ parse_test.go
@@ -1,11 +1,58 @@
 package dicom_test
 
 import (
+	"encoding/json"
+	"fmt"
+	"github.com/suyashkumar/dicom/pkg/frame"
+	"github.com/suyashkumar/dicom/pkg/tag"
+	"image/jpeg"
+	"os"
 	"testing"
 
 	"github.com/suyashkumar/dicom"
 	"github.com/suyashkumar/dicom/pkg/dcmtest"
 )
+
+func Example_readFile() {
+	// See also: dicom.Parse, which uses a more generic io.Reader API.
+	dataset, _ := dicom.ParseFile("./pkg/dcmtest/data/1.dcm", nil)
+
+	// Dataset will nicely print the DICOM dataset data out of the box.
+	fmt.Println(dataset)
+
+	// Dataset is also JSON serializable out of the box.
+	j, _ := json.Marshal(dataset)
+	fmt.Println(j)
+}
+
+func Example_streamingFrames() {
+	frameChan := make(chan *frame.Frame)
+
+	// Go routine to handle streaming frames as they are parsed. This may be
+	// useful when parsing a large DICOM with many frames from a network source,
+	// where image frames can start to be processed before the entire DICOM
+	// is parsed (or even read from storage).
+	go func() {
+		for fr := range frameChan {
+			fmt.Println(fr)
+		}
+	}()
+
+	dataset, _ := dicom.ParseFile("./pkg/dcmtest/data/1.dcm", frameChan)
+	fmt.Println(dataset) // The full dataset
+}
+
+func Example_getImageFrames() {
+	dataset, _ := dicom.ParseFile("./pkg/dcmtest/data/1.dcm", nil)
+	pixelDataElement, _ := dataset.FindElementByTag(tag.PixelData)
+	pixelDataInfo := dicom.MustGetPixelDataInfo(pixelDataElement.Value)
+	for i, fr := range pixelDataInfo.Frames {
+		img, _ := fr.GetImage() // The Go image.Image for this frame
+		f, _ := os.Create(fmt.Sprintf("image_%d.jpg", i))
+		_ = jpeg.Encode(f, img, &jpeg.Options{Quality: 100})
+		_ = f.Close()
+	}
+}
 
 // TestParse is an end-to-end sanity check over DICOMs in testdata/. Currently it only checks that no error is returned
 // when parsing the files.
